% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_classes}
\alias{two_classes}
\title{Measurements of sequence recognition from yeast proteins with different subcellular localisations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35 rows and 2 columns.}
\usage{
two_classes
}
\description{
This is a subset of the \link{yeast} data set. See the\link{yeast} help page for further info
}
\keyword{datasets}
